# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Kavach Server API",
        "title": "Kavach Server API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/",
    "paths": {
...
        "/organisations/{organisation_id}/tokens": {
            "get": {
                "description": "Show a organisation tokens",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrganisationTokens"
                ],
                "summary": "List organisation tokens",
                "operationId": "get-organisation-tokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.OrganisationToken"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create token for an organisation using organisation_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OrganisationTokens"
                ],
                "summary": "Create token for an organisation using organisation_id",
                "operationId": "create-organisation-token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "X-User",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Organisation ID",
                        "name": "organisation_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Organisation Token Body",
                        "name": "OrganisationTokenBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OrganisationToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OrganisationToken"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Application": {
            "type": "object",
            "properties": {
                "application_role": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ApplicationRole"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "medium_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation": {
                    "$ref": "#/definitions/model.Organisation"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "spaces": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Space"
                    }
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ApplicationToken"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.ApplicationRole": {
            "type": "object",
            "properties": {
                "application": {
                    "$ref": "#/definitions/model.Application"
                },
                "application_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.ApplicationToken": {
            "type": "object",
            "properties": {
                "application": {
                    "$ref": "#/definitions/model.Application"
                },
                "application_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.Medium": {
            "type": "object",
            "properties": {
                "alt_text": {
                    "type": "string"
                },
                "caption": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dimensions": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.Organisation": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Application"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "organisation_users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrganisationUser"
                    }
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OrganisationRole"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.OrganisationRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation": {
                    "$ref": "#/definitions/model.Organisation"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.OrganisationToken": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation": {
                    "$ref": "#/definitions/model.Organisation"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
        "model.OrganisationUser": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "organisation": {
                    "$ref": "#/definitions/model.Organisation"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Space": {
            "type": "object",
            "properties": {
                "analytics": {
                    "type": "string"
                },
                "application": {
                    "$ref": "#/definitions/model.Application"
                },
                "application_id": {
                    "type": "integer"
                },
                "contact_info": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fav_icon": {
                    "$ref": "#/definitions/model.Medium"
                },
                "fav_icon_id": {
                    "type": "integer"
                },
                "footer_code": {
                    "type": "string"
                },
                "header_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "$ref": "#/definitions/model.Medium"
                },
                "logo_id": {
                    "type": "integer"
                },
                "logo_mobile": {
                    "$ref": "#/definitions/model.Medium"
                },
                "logo_mobile_id": {
                    "type": "integer"
                },
                "meta_fields": {
                    "type": "string"
                },
                "mobile_icon": {
                    "$ref": "#/definitions/model.Medium"
                },
                "mobile_icon_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organisation": {
                    "$ref": "#/definitions/model.Organisation"
                },
                "organisation_id": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SpaceRole"
                    }
                },
                "site_address": {
                    "type": "string"
                },
                "site_title": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "social_media_urls": {
                    "type": "string"
                },
                "tag_line": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                },
                "verification_codes": {
                    "type": "string"
                }
            }
        },
        "model.SpaceRole": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "space": {
                    "$ref": "#/definitions/model.Space"
                },
                "space_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "featured_medium_id": {
                    "type": "integer"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kid": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "medium": {
                    "$ref": "#/definitions/model.Medium"
                },
                "organisations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Organisation"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "social_media_urls": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}